requirements.txt
praw~=7.5.0
python-decouple==3.5